fix: restructure to allow testing
