+++ test/setup.ts
@@ -3,6 +3,8 @@ import { vi } from 'vitest'
 // setup global constants
 vi.stubGlobal('__I18N_CACHE__', false)
 vi.stubGlobal('__I18N_CACHE_LIFETIME__', -1)
+vi.stubGlobal('__I18N_STRICT_SEO__', false)
+vi.stubGlobal('__PARALLEL_PLUGIN__', false)
 
 vi.stubGlobal('__TRAILING_SLASH__', false)
 vi.stubGlobal('__DIFFERENT_DOMAINS__', false)
